+++ InventoryManagement.Domain/Interfaces/Repositories/IProductRepository.cs
@@ -1,9 +1,4 @@
 ﻿using InventoryManagement.Domain.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace InventoryManagement.Domain.Interfaces.Repositories
 {
@@ -12,11 +7,11 @@ namespace InventoryManagement.Domain.Interfaces.Repositories
     /// </summary>
     public interface IProductRepository : IRepository<Product>
     {
-        Task<Product> GetProductWithInventoryAsync(int productId);
+        Task<Product> GetProductWithInventoryAsync(long productId);
         Task<Product> GetProductBySKUAsync(string sku);
         Task<IReadOnlyList<Product>> GetProductsByCategoryAsync(string category);
-        Task<IReadOnlyList<Product>> GetLowStockProductsAsync(int threshold);
-        Task<bool> IsSkuUniqueAsync(string sku, int? excludeProductId = null);
+        Task<IReadOnlyList<Product>> GetLowStockProductsAsync(long threshold);
+        Task<bool> IsSkuUniqueAsync(string sku, long? excludeProductId = null);
         Task<Dictionary<string, int>> GetProductCountByCategoryAsync();
     }
 }
